added dots for each post type and restructured some code in app comp
